add ShopConstants file
